requirements.txt
Django==3.0.4
django-bootstrap-modal-forms==1.5.0
django-widget-tweaks==1.4.5
